Reflect the current state of this for on README.md
Sadly the latest activity on `master` was over 10 months ago. Saying that "I will maintain this repo and I will accept pull requests" is a bit misleading and it might encourage people looking for an active fork to adapt it just to regret it deeply afterwards. It happened with me, at least.

I hope this pull request (eventually) gets merged.